clean up cli.py and move all command methods to commands directory
