--- src/silvimetric/commands/__init__.py
@@ -1,3 +1,0 @@
-from .scan import scan
-from .extract import extract
-from .shatter import shatter
